Add files via upload
The EY NextWave Data Science Challenge 2019 focuses on how data can help the next smart city thrive, and boost the mobility of the future. Global urbanization is on the rise, with more than 50% of the world’s population living in cities; according to the UN, that number will reach 60% by 2030 – that’s nearly 1.5 billion more than in 2010.